Переработан фрагмент заказов + обновлены стили кнопок
