+++ partner-app/src/main/java/ru/anykeyers/partner_app/ui/fragment/account/company/CompanyFragment.kt
@@ -42,6 +42,10 @@ class CompanyFragment : Fragment() {
             config = it
         }
 
+        binding.reportButton.setOnClickListener {
+            vm.loadReport()
+        }
+
         binding.submitButton.setOnClickListener {
             val info = OrganizationInfo(
                 config.organizationInfo.tin,
