+++ partner-app/src/main/java/ru/anykeyers/partner_app/ui/vm/AccountViewModel.kt
@@ -1,12 +1,7 @@
 package ru.anykeyers.partner_app.ui.vm
 
-import android.app.DownloadManager
-import android.content.Context
-import android.net.Uri
-import android.os.Environment
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import ru.anykeyers.partner_app.data.remote.WebConstant
 import ru.anykeyers.partner_app.domain.entity.User
 import ru.anykeyers.partner_app.domain.repository.IUserRepository
 
@@ -15,7 +10,6 @@ import ru.anykeyers.partner_app.domain.repository.IUserRepository
  */
 class AccountViewModel(
     private val userRepository: IUserRepository,
-    private val context: Context
 ) : HandlingViewModel() {
 
     private val _user by lazy { MutableLiveData<User>() }
@@ -32,39 +26,6 @@ class AccountViewModel(
     fun updateUser(user: User) {
         launchWithResultState {
             userRepository.updateUser(user)
-        }
-    }
-
-    /**
-     * Загрузка отчета автомойки в формате PDF
-     */
-    fun loadReport() {
-        val userId = _user.value?.id ?: return
-        val reportUrl = "${WebConstant.CAR_WASH_SERVICE_URL}/configuration/pdf/$userId"
-        launchWithResultState {
-            downloadPdfFile(reportUrl)
-        }
-    }
-
-    /**
-     * Скачивание PDF-файла по указанному URL
-     */
-    private fun downloadPdfFile(url: String) {
-        val request = createDownloadRequest(url)
-        val downloadManager = context.getSystemService(Context.DOWNLOAD_SERVICE) as? DownloadManager
-        downloadManager?.enqueue(request)
-    }
-
-    /**
-     * Создает запрос на загрузку файла
-     */
-    private fun createDownloadRequest(url: String): DownloadManager.Request {
-        return DownloadManager.Request(Uri.parse(url)).apply {
-            setTitle("Отчет автомойки")
-            setDescription("Идёт загрузка PDF-файла")
-            setNotificationVisibility(DownloadManager.Request.VISIBILITY_VISIBLE_NOTIFY_COMPLETED)
-            setDestinationInExternalPublicDir(Environment.DIRECTORY_DOWNLOADS, "report.pdf")
-            setMimeType("application/pdf")
         }
     }
