+++ partner-app/src/main/res/layout/fragment_account.xml
@@ -71,7 +71,7 @@
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:orientation="vertical"
-                    android:layout_marginBottom="24dp">
+                    android:layout_marginBottom="18dp">
 
                     <TextView
                         android:layout_width="match_parent"
@@ -80,7 +80,8 @@
                         android:textSize="18sp"
                         android:textColor="@android:color/black"
                         android:fontFamily="sans-serif-medium"
-                        android:layout_marginBottom="8dp" />
+                        android:textStyle="bold"
+                        android:layout_marginBottom="12dp" />
 
                     <LinearLayout
                         android:layout_width="match_parent"
@@ -95,9 +96,9 @@
                             android:layout_height="wrap_content"
                             android:text="@string/profile.notification.self"
                             android:textSize="16sp"
-                            android:textColor="@android:color/darker_gray"
+                            android:textColor="@android:color/black"
                             android:fontFamily="sans-serif"
-                            android:layout_marginBottom="4dp" />
+                            android:layout_marginBottom="12dp" />
 
                         <TextView
                             android:id="@+id/notificationSettingsTextView"
@@ -105,7 +106,7 @@
                             android:layout_height="wrap_content"
                             android:text="@string/profile.notification.settings"
                             android:textSize="16sp"
-                            android:textColor="@android:color/darker_gray"
+                            android:textColor="@android:color/black"
                             android:fontFamily="sans-serif" />
                     </LinearLayout>
                 </LinearLayout>
@@ -120,9 +121,10 @@
                         android:layout_height="wrap_content"
                         android:text="@string/profile.management.header"
                         android:textSize="18sp"
+                        android:textStyle="bold"
                         android:textColor="@android:color/black"
                         android:fontFamily="sans-serif-medium"
-                        android:layout_marginBottom="8dp" />
+                        android:layout_marginBottom="12dp" />
 
                     <LinearLayout
                         android:layout_width="match_parent"
@@ -137,9 +139,9 @@
                             android:layout_height="wrap_content"
                             android:text="@string/profile.management.comp_info"
                             android:textSize="16sp"
-                            android:textColor="@android:color/darker_gray"
+                            android:textColor="@android:color/black"
                             android:fontFamily="sans-serif"
-                            android:layout_marginBottom="4dp" />
+                            android:layout_marginBottom="12dp" />
 
                         <TextView
                             android:id="@+id/employees"
@@ -147,19 +149,11 @@
                             android:layout_height="wrap_content"
                             android:text="@string/profile.management.employees"
                             android:textSize="16sp"
-                            android:textColor="@android:color/darker_gray"
+                            android:textColor="@android:color/black"
                             android:fontFamily="sans-serif"
                             android:layout_marginBottom="4dp" />
                     </LinearLayout>
                 </LinearLayout>
-
-                <Button
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="16dp"
-                    android:id="@+id/report_button"
-                    style="@style/PrimaryButton"
-                    android:text="@string/profile.report.generate"/>
             </LinearLayout>
         </androidx.cardview.widget.CardView>
     </LinearLayout>
