+++ partner-app/src/main/res/layout/fragment_box_create.xml
@@ -51,12 +51,9 @@
                     android:orientation="horizontal">
 
                     <Button
+                        style="@style/SecondaryButton"
                         android:id="@+id/delete_button"
-                        style="@style/AdditionalButton"
-                        android:layout_width="0dp"
-                        android:layout_height="wrap_content"
                         android:layout_marginEnd="8dp"
-                        android:layout_weight="1"
                         android:text="@string/common.delete" />
 
                     <Button
@@ -71,7 +68,7 @@
                 <Button
                     android:id="@+id/submit_button"
                     style="@style/PrimaryButton"
-                    android:text="@string/common.save" />
+                    android:text="@string/common.add" />
             </LinearLayout>
         </androidx.cardview.widget.CardView>
     </LinearLayout>
