+++ partner-app/src/main/res/layout/fragment_company.xml
@@ -121,10 +121,23 @@
                         android:padding="12dp" />
                 </LinearLayout>
 
-                <Button
-                    android:id="@+id/submit_button"
-                    style="@style/PrimaryButton"
-                    android:text="@string/common.save"/>
+                <LinearLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:orientation="vertical">
+
+                    <Button
+                        style="@style/SecondaryButton"
+                        android:layout_width="match_parent"
+                        android:id="@+id/report_button"
+                        android:layout_marginBottom="8dp"
+                        android:text="@string/profile.report.generate" />
+
+                    <Button
+                        android:id="@+id/submit_button"
+                        style="@style/PrimaryButton"
+                        android:text="@string/common.save"/>
+                </LinearLayout>
             </LinearLayout>
         </androidx.cardview.widget.CardView>
     </LinearLayout>
