+++ partner-app/src/main/res/layout/fragment_home.xml
@@ -202,12 +202,14 @@
                                     android:id="@+id/inProcess"
                                     android:layout_width="wrap_content"
                                     android:layout_height="wrap_content"
+                                    android:textColor="@color/black"
                                     android:text="@string/home.order.running"
                                     android:textSize="14sp"/>
 
                                 <TextView
                                     android:id="@+id/inProcessCount"
                                     android:layout_width="wrap_content"
+                                    android:textColor="@color/black"
                                     android:layout_height="wrap_content"
                                     android:text="1"/>
                             </LinearLayout>
@@ -244,12 +246,14 @@
                                     android:id="@+id/inQueue"
                                     android:layout_width="wrap_content"
                                     android:layout_height="wrap_content"
+                                    android:textColor="@color/black"
                                     android:text="@string/home.order.wait_process"
                                     android:textSize="14sp"/>
 
                                 <TextView
                                     android:id="@+id/inQueueCount"
                                     android:layout_width="wrap_content"
+                                    android:textColor="@color/black"
                                     android:layout_height="wrap_content"
                                     android:text="4"/>
                             </LinearLayout>
@@ -286,6 +290,7 @@
                                     android:id="@+id/processing"
                                     android:layout_width="wrap_content"
                                     android:layout_height="wrap_content"
+                                    android:textColor="@color/black"
                                     android:text="@string/home.order.processing"
                                     android:textSize="14sp"/>
 
@@ -293,6 +298,7 @@
                                     android:id="@+id/processingCount"
                                     android:layout_width="wrap_content"
                                     android:layout_height="wrap_content"
+                                    android:textColor="@color/black"
                                     android:text="1"/>
                             </LinearLayout>
                         </androidx.cardview.widget.CardView>
@@ -328,6 +334,7 @@
                                     android:id="@+id/completed"
                                     android:layout_width="wrap_content"
                                     android:layout_height="wrap_content"
+                                    android:textColor="@color/black"
                                     android:text="@string/home.order.processed"
                                     android:textSize="14sp"/>
 
@@ -335,6 +342,7 @@
                                     android:id="@+id/completedCount"
                                     android:layout_width="wrap_content"
                                     android:layout_height="wrap_content"
+                                    android:textColor="@color/black"
                                     android:text="8"/>
                             </LinearLayout>
                         </androidx.cardview.widget.CardView>
