+++ partner-app/src/main/res/layout/fragment_orders.xml
@@ -45,6 +45,7 @@
                             app:layout_constraintStart_toEndOf="@id/order_fragment_title"
                             app:layout_constraintTop_toTopOf="@+id/order_fragment_title"
                             app:layout_constraintBottom_toBottomOf="@+id/order_fragment_title"
+                            android:layout_marginStart="8dp"
                             android:layout_marginEnd="8dp"
                             tools:ignore="ContentDescription" />
