+++ partner-app/src/main/res/layout/fragment_service.xml
@@ -23,5 +23,4 @@
             android:layout_height="wrap_content"
             android:layout_marginTop="16dp" />
     </LinearLayout>
-
 </androidx.constraintlayout.widget.ConstraintLayout>
