+++ partner-app/src/main/res/layout/fragment_service_create.xml
@@ -75,17 +75,15 @@
                     android:orientation="horizontal">
 
                     <Button
+                        style="@style/SecondaryButton"
                         android:id="@+id/delete_button"
-                        style="@style/AdditionalButton"
-                        android:layout_width="0dp"
                         android:layout_height="wrap_content"
                         android:layout_marginEnd="8dp"
-                        android:layout_weight="1"
                         android:text="@string/common.delete" />
 
                     <Button
-                        android:id="@+id/update_button"
                         style="@style/PrimaryButton"
+                        android:id="@+id/update_button"
                         android:layout_width="0dp"
                         android:layout_height="wrap_content"
                         android:layout_weight="1"
@@ -93,12 +91,10 @@
                 </LinearLayout>
 
                 <Button
-                    android:id="@+id/submit_button"
                     style="@style/PrimaryButton"
-                    android:text="@string/common.save" />
+                    android:id="@+id/submit_button"
+                    android:text="@string/common.add" />
             </LinearLayout>
-
         </androidx.cardview.widget.CardView>
     </LinearLayout>
-
 </androidx.constraintlayout.widget.ConstraintLayout>
