+++ partner-app/src/main/res/values/styles.xml
@@ -9,20 +9,29 @@
         <item name="android:layout_height">wrap_content</item>
     </style>
 
-    <style name="AdditionalButton" parent="Widget.MaterialComponents.Button">
+    <style name="SecondaryButton" parent="Widget.AppCompat.Button">
         <item name="android:backgroundTint">@color/blue_50</item>
         <item name="android:background">@drawable/rounded_button_background</item>
-        <item name="android:textColor">@color/blue_50</item>
         <item name="android:textSize">8pt</item>
-        <item name="android:layout_width">match_parent</item>
+        <item name="android:textAllCaps">false</item>
+        <item name="android:textColor">@color/blue_500</item>
+        <item name="android:elevation">0dp</item>
+        <item name="android:layout_width">0dp</item>
+        <item name="android:layout_weight">1</item>
         <item name="android:layout_height">wrap_content</item>
+    </style>
+
+    <style name="PrimaryItemCard" parent="CardView">
+        <item name="cardCornerRadius">18dp</item>
+        <item name="cardElevation">0dp</item>
+        <item name="cardBackgroundColor">@color/blue_50</item>
     </style>
 
     <style name="InfoCard" parent="Widget.MaterialComponents.CardView">
         <item name="android:layout_width">match_parent</item>
         <item name="android:layout_height">wrap_content</item>
         <item name="cardCornerRadius">28dp</item>
-        <item name="cardElevation">4dp</item>
+        <item name="cardElevation">0dp</item>
         <item name="cardBackgroundColor">@android:color/white</item>
     </style>
